Adding basic.convert user documentation

+++ molsysmt/item/pytraj_Topology/to_pytraj_Trajectory.py
@@ -1,5 +1,6 @@
 from molsysmt._private.digestion import digest
 from molsysmt._private.exceptions import LibraryNotFoundError
+import numpy as np
 
 @digest(form='pytraj.Topology')
 def to_pytraj_Trajectory(item, atom_indices='all', coordinates=None, box=None):
@@ -11,12 +12,12 @@ def to_pytraj_Trajectory(item, atom_indices='all', coordinates=None, box=None):
 
     from . import extract
     from molsysmt import pyunitwizard as puw
-    from molsysmt.pbc import box_to_box_angles
+    from molsysmt.pbc import box_angles_from_box, box_lengths_from_box
 
     tmp_item = extract(item, atom_indices=atom_indices, copy_if_all=False)
 
     coordinates = puw.get_value(coordinates, to_unit='angstroms')
-    tmp_item = pt.Trajectory(xyz=coordinates, top=tmp_item)
+    tmp_item = pt.Trajectory(xyz=coordinates.astype('float64'), top=tmp_item)
 
     box_lengths = box_lengths_from_box(box)
     box_angles = box_angles_from_box(box)
@@ -25,6 +26,7 @@ def to_pytraj_Trajectory(item, atom_indices='all', coordinates=None, box=None):
     box_angles = puw.get_value(box_angles, to_unit='degrees')
 
     tmp_item.unitcells = np.hstack([box_lengths, box_angles])
+    tmp_item.unitcells = tmp_item.unitcells.astype('float64')
 
     return tmp_item
